ocr-service/requirements.txt
# FastAPI
fastapi
uvicorn[standard]

# For the OCR
pytesseract

# For downloading and opening the image
Pillow
httpx